# Modify Teddy's match verification.