@BurntSushi 2016-09-04
I rebased, squashed and merged this in https://github.com/rust-lang-nursery/regex/commit/12cb63bfc16440ad23661f68d2a22930897d8c05

Thanks!