@BurntSushi 2016-09-05
@jneem I somehow missed your comment about AVX2. Are you working on that? That'd be delicious. :-)