@jneem 2016-09-06
Yes, but it's in a separate crate (https://github.com/jneem/teddy), which has diverged a bit. It also implements aligned loads, which gives another 10% or so.